+++ crates/translator_new_rml/src/rml_model/v2/core/expression_map/error.rs
@@ -1,0 +1,7 @@
+use thiserror::Error;
+
+#[derive(Debug, Error)]
+pub enum ExpressionMapError {
+    #[error("invalid template: `{0}` : {1}")]
+    InvalidTemplate(String, String),
+}